fix(reviews): moved folders to the right place
